# swagger.yaml
openapi: 3.0.3
info:
  title: E-Commerce API — Testing Practice
  version: 1.0.0
  description: |
    Full-featured e-commerce REST API built for API testing practice.
    Covers auth flows, CRUD, pagination (offset + cursor), caching (ETag/304),
    file uploads, BOLA/IDOR, rate limiting, fault simulation, idempotency,
    RFC 7807 errors, state machines, and role-based access.

    ## Seeded credentials (reset anytime via `POST /v1/seed/reset`)
    | Role  | Email             | Password   |
    |-------|-------------------|------------|
    | admin | admin@test.com    | Admin123!  |
    | user  | user@test.com     | User123!   |
    | user  | user2@test.com    | User123!   |

    ## Notable testing hooks
    - `?_delay=3000` — delays response by N ms (max 10 000)
    - `?_fail=503`   — returns a simulated error with that status code
    - `Idempotency-Key` header on `POST /orders` — prevents duplicate orders on retry
    - `If-None-Match` header on GET product(s) — returns 304 when data unchanged
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://YOUR-APP.onrender.com/v1
    description: Production (update after deploying to Render)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details
      required: [type, title, status, detail, instance]
      properties:
        type:     { type: string, example: "http://localhost:3000/errors/not-found" }
        title:    { type: string, example: "Not Found" }
        status:   { type: integer, example: 404 }
        detail:   { type: string, example: "Product with id 99 does not exist" }
        instance: { type: string, example: "/v1/products/99" }
...
    OrderItem:
      type: object
      properties:
        productId:   { type: integer }
        productName: { type: string }
        price:       { type: number }
        quantity:    { type: integer }
        subtotal:    { type: number }
    Order:
      type: object
      properties:
        id:        { type: integer, example: 1 }
        userId:    { type: integer, example: 2 }
        items:     { type: array, items: { $ref: '#/components/schemas/OrderItem' } }
        total:     { type: number, example: 249.99 }
        status:    { type: string, enum: [pending, confirmed, shipped, delivered, cancelled] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OffsetPagination:
      type: object
      properties:
        type:       { type: string, example: offset }
        page:       { type: integer }
        limit:      { type: integer }
        total:      { type: integer }
        totalPages: { type: integer }
        hasNext:    { type: boolean }
        hasPrev:    { type: boolean }
    CursorPagination:
      type: object
      properties:
        type:       { type: string, example: cursor }
        limit:      { type: integer }
        nextCursor: { type: integer, nullable: true }
        hasNext:    { type: boolean }
  responses:
    Unauthorized:
      description: Unauthorized — missing, expired, or invalid token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Forbidden — insufficient role
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
paths:
...
  /orders:
    post:
      summary: Checkout — converts cart to order
      description: |
        Validates stock, snapshots prices, decrements stock, clears cart.
        Supports `Idempotency-Key` header — sending same key twice returns same 201 (no duplicate order).
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string }
          description: Unique key per checkout attempt — enables safe retry
      responses:
        '201':
          headers:
            Location: { schema: { type: string } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } }
        '400': { description: Cart is empty }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { description: Insufficient stock or product unavailable }
    get:
      summary: List orders (user sees own; admin sees all)
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - { name: page,   in: query, schema: { type: integer } }
        - { name: limit,  in: query, schema: { type: integer } }
        - { name: cursor, in: query, schema: { type: integer }, description: Cursor-based pagination }
        - { name: status, in: query, schema: { type: string, enum: [pending, confirmed, shipped, delivered, cancelled] } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: '#/components/schemas/Order' } }
                  pagination:
                    oneOf:
                      - { $ref: '#/components/schemas/OffsetPagination' }
                      - { $ref: '#/components/schemas/CursorPagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /orders/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      summary: Get order details
      description: Users can only access their own orders (BOLA/IDOR protection). Use `?_delay` / `?_fail` for fault simulation.
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - { name: _delay, in: query, schema: { type: integer } }
        - { name: _fail,  in: query, schema: { type: integer } }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orders/{id}/status:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    patch:
      summary: Update order status (admin — state machine enforced)
      description: |
        Valid transitions:
        `pending → confirmed | cancelled`
        `confirmed → shipped | cancelled`
        `shipped → delivered`
        `delivered / cancelled → (terminal, no transitions)`
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [confirmed, shipped, delivered, cancelled] }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: Invalid state transition }
  /orders/{id}/cancel:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    post:
      summary: Cancel order (owner or admin)
      description: Returns 409 if order is shipped or delivered.
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Cannot cancel — order already shipped or delivered }

  # ── Users ───────────────────────────────────────────────────────────────────
...